Add : 헤더 푸터 연결, scss풀리는 현상 해결
