+++ src/components/CategoryModal/CategoryModalNav.js
@@ -52,6 +52,7 @@ const CategoryModalNav = ({
           onMouseLeave={() => setSelectedMainCategory(null)}
         >
           <Link
+            style={{ textDecoration: 'none' }}
             to={`/productlist?categoryId=${mainCategory.category_id}&limit=9&orderBy=&offset=0`}
           >
             <img
@@ -77,6 +78,7 @@ const CategoryModalNav = ({
                     key={subCategory.sub_category_id}
                   >
                     <Link
+                      style={{ textDecoration: 'none' }}
                       to={`/productlist?categoryId=${subCategory.category_id}&subCategoryId=${subCategory.subCategory_id}&limit=9&orderBy=&offset=0`}
                     >
                       <img
