+++ src/components/Footer/Footer.scss
@@ -1,11 +1,22 @@
 .footer {
   height: auto;
+  margin-top: 90px;
   border-top: 1px solid rgb(221, 221, 221);
 
   &-list {
     display: flex;
     padding: 48px 0 24px 16px;
     gap: 48px;
+    .to-login {
+      border: none;
+      outline: none;
+      background-color: transparent;
+      color: #999999;
+
+      &:hover {
+        cursor: pointer;
+      }
+    }
   }
 
   &-item {
