+++ src/pages/Details/Mainlayout.js
@@ -1,16 +1,45 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
+import { useLocation, useNavigate } from 'react-router-dom';
 import Footer from '../../components/Footer/Footer';
 import NavMain from '../../components/Nav/NavMain';
-import TitleLine from '../../components/TitleLine/TitleLine';
+import Header from '../../components/Header/Header';
 
 const MainLayout = props => {
   const { children } = props;
+  const [isLoggedIn, setIsLoggedIn] = useState(false);
+  const location = useLocation();
+  const navigate = useNavigate();
+  const handleLogin = () => {
+    navigate('/login');
+  };
+
+  useEffect(() => {
+    const token = localStorage.getItem('token');
+    setIsLoggedIn(!!token); // 토큰이 있으면 true, 없으면 false
+  }, [location.pathname, setIsLoggedIn]);
+
+  const handleButtonClick = () => {
+    if (isLoggedIn) {
+      localStorage.removeItem('token');
+      setIsLoggedIn(false);
+    } else {
+      navigate('/login');
+    }
+  };
   return (
     <div style={{ display: 'flex', flexDirection: 'column', height: '100vh' }}>
-      <TitleLine />
+      <Header
+        isLoggedIn={isLoggedIn}
+        handleLogin={handleLogin}
+        handleButtonClick={handleButtonClick}
+      />
       <NavMain />
-      <div style={{ flex: '1', marginTop: '48px' }}>{children}</div>
-      <Footer />
+      <div style={{ flex: '1' }}>{children}</div>
+      <Footer
+        isLoggedIn={isLoggedIn}
+        handleLogin={handleLogin}
+        handleButtonClick={handleButtonClick}
+      />
     </div>
   );
 };
