+++ src/pages/Main/Main.scss
@@ -1,6 +1,8 @@
 @import '../../styles/variables';
 
 .main {
+  display: flex;
+  flex-direction: column;
   a {
     color: #1c1c1c;
     text-decoration: none;
