+++ src/pages/Main/components/NewestNav.scss
@@ -2,12 +2,12 @@
 
 .newest-nav {
   width: 100%;
-  height: 360px;
 
   .newest-navbar {
     display: flex;
     align-items: center;
     justify-content: space-around;
+    height: 50px;
     margin-top: 40px;
   }
