+++ src/pages/Main/components/NewestPrdouct.scss
@@ -6,9 +6,10 @@
   align-items: flex-start !important;
   justify-content: flex-start !important;
   width: 1050px;
+
   margin-top: 32px;
   gap: 70px;
- 
+
   .product-item {
     width: 200px;
     transform: translateZ(0);
@@ -37,7 +38,7 @@
     .product-price {
       @include font(18px, #1c1c1c);
     }
-  
+
     .to-cart {
       position: absolute;
       right: 0px;
